refactor(projects): restructure PluginDemo and Loader into parent/child layout with facades
Reorganize ATAP.Console.PluginDemo into facade + Model + StringConstants subprojects.
Reorganize ATAP.Utilities.Loader into facade + Model + Interfaces + StringConstants subprojects.
Add FodyWeavers.xml/xsd for PluginDemo and Model.
Add EventSource attribute to ATAPUtilitiesETWProvider and reformat.
Add TRACE to Debug DefineConstants in Directory.Build.props.
Update solution file and all dependent .csproj project references.
Add Install-SqlServerInstance.ps1 and update Build-DatabaseWithFlyway.ps1.
