+++ ATAP.Utilities.sln
@@ -378,13 +378,13 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ATAP.Utilities.Serializer.S
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ATAP.Utilities.Serializer.StringConstants", "src\ATAP.Utilities.Serializer\StringConstants\ATAP.Utilities.Serializer.StringConstants.csproj", "{4DDA9B1A-ADA9-4BD9-86CC-D69F1BE14416}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ATAP.Utilities.Loader.StringConstants", "src\ATAP.Utilities.Loader.StringConstants\ATAP.Utilities.Loader.StringConstants.csproj", "{35B1ACB4-D905-40AE-A113-46A3307B0686}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ATAP.Utilities.Loader.StringConstants", "src\ATAP.Utilities.Loader\StringConstants\ATAP.Utilities.Loader.StringConstants.csproj", "{35B1ACB4-D905-40AE-A113-46A3307B0686}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ATAP.Utilities.Loader.Interfaces", "src\ATAP.Utilities.Loader.Interfaces\ATAP.Utilities.Loader.Interfaces.csproj", "{8DB16880-B7AB-4C44-A6E3-BE1D7CFA3F43}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ATAP.Utilities.Loader.Interfaces", "src\ATAP.Utilities.Loader\Interfaces\ATAP.Utilities.Loader.Interfaces.csproj", "{8DB16880-B7AB-4C44-A6E3-BE1D7CFA3F43}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ATAP.Utilities.Loader", "src\ATAP.Utilities.Loader\ATAP.Utilities.Loader.csproj", "{955751DB-DDA8-4131-A437-98D2D2F86E41}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ATAP.Utilities.Loader.Model", "src\ATAP.Utilities.Loader\Model\ATAP.Utilities.Loader.Model.csproj", "{955751DB-DDA8-4131-A437-98D2D2F86E41}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Loader", "Loader", "{08E1C810-578A-11EB-8BEC-313999507D9E}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ATAP.Utilities.Loader", "src\ATAP.Utilities.Loader\ATAP.Utilities.Loader.csproj", "{8CDA64FB-B6EA-4045-B141-0D02652F22B9}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ATAP.Utilities.Philote.JsonConverter.Shim.SystemTextJson", "src\ATAP.Utilities.Philote\JsonConverter.Shim.SystemTextJson\ATAP.Utilities.Philote.JsonConverter.Shim.SystemTextJson.csproj", "{B11FF8C5-60FD-4AEA-9021-0861D8F820FC}"
 EndProject
@@ -1148,6 +1148,12 @@ Global
 		{955751DB-DDA8-4131-A437-98D2D2F86E41}.Release|Any CPU.Build.0 = Release|Any CPU
 		{955751DB-DDA8-4131-A437-98D2D2F86E41}.ReleaseWithTrace|Any CPU.ActiveCfg = ReleaseWithTrace|Any CPU
 		{955751DB-DDA8-4131-A437-98D2D2F86E41}.ReleaseWithTrace|Any CPU.Build.0 = ReleaseWithTrace|Any CPU
+		{8CDA64FB-B6EA-4045-B141-0D02652F22B9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8CDA64FB-B6EA-4045-B141-0D02652F22B9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8CDA64FB-B6EA-4045-B141-0D02652F22B9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8CDA64FB-B6EA-4045-B141-0D02652F22B9}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8CDA64FB-B6EA-4045-B141-0D02652F22B9}.ReleaseWithTrace|Any CPU.ActiveCfg = ReleaseWithTrace|Any CPU
+		{8CDA64FB-B6EA-4045-B141-0D02652F22B9}.ReleaseWithTrace|Any CPU.Build.0 = ReleaseWithTrace|Any CPU
 		{B11FF8C5-60FD-4AEA-9021-0861D8F820FC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{B11FF8C5-60FD-4AEA-9021-0861D8F820FC}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B11FF8C5-60FD-4AEA-9021-0861D8F820FC}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -1592,10 +1598,10 @@ Global
 		{DA3CA34F-68BB-4B3C-86B3-720B2FE225C5} = {6CA48A6B-4910-4DD7-92D1-F079CD34A13B}
 		{C447A64E-FBE5-4420-B698-EC6EC3A42636} = {DC413220-33A6-11EB-9C14-5132EBA7B73D}
 		{4DDA9B1A-ADA9-4BD9-86CC-D69F1BE14416} = {DC413220-33A6-11EB-9C14-5132EBA7B73D}
-		{35B1ACB4-D905-40AE-A113-46A3307B0686} = {08E1C810-578A-11EB-8BEC-313999507D9E}
-		{8DB16880-B7AB-4C44-A6E3-BE1D7CFA3F43} = {08E1C810-578A-11EB-8BEC-313999507D9E}
-		{955751DB-DDA8-4131-A437-98D2D2F86E41} = {08E1C810-578A-11EB-8BEC-313999507D9E}
-		{08E1C810-578A-11EB-8BEC-313999507D9E} = {F78E1157-C4C3-4162-8B2D-72FC45CB698E}
+		{35B1ACB4-D905-40AE-A113-46A3307B0686} = {8CDA64FB-B6EA-4045-B141-0D02652F22B9}
+		{8DB16880-B7AB-4C44-A6E3-BE1D7CFA3F43} = {8CDA64FB-B6EA-4045-B141-0D02652F22B9}
+		{955751DB-DDA8-4131-A437-98D2D2F86E41} = {8CDA64FB-B6EA-4045-B141-0D02652F22B9}
+		{8CDA64FB-B6EA-4045-B141-0D02652F22B9} = {F78E1157-C4C3-4162-8B2D-72FC45CB698E}
 		{B11FF8C5-60FD-4AEA-9021-0861D8F820FC} = {D27866D0-340C-11EB-B0F3-8BBCA7824458}
 		{B6032930-1FA8-11EC-9C72-27E1FBAA6FB1} = {F78E1157-C4C3-4162-8B2D-72FC45CB698E}
 		{FB2917AC-BDCB-4012-9130-E2EDA9CA7899} = {B6032930-1FA8-11EC-9C72-27E1FBAA6FB1}
