+++ src/ATAP.Console.PluginDemo/ATAP.Console.PluginDemo.csproj
@@ -1,56 +1,19 @@
 <Project Sdk="Microsoft.NET.Sdk">
-
   <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFrameworks>net10.0</TargetFrameworks>
-    <RootNamespace>ATAP.Console.PluginDemo</RootNamespace>
-    <GeneratePackageOnBuild>false</GeneratePackageOnBuild>
-    <IsPackable>false</IsPackable>
+    <OutputType>Library</OutputType>
+    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
+    <IsPackable>true</IsPackable>
     <MajorVersion>0</MajorVersion>
     <MinorVersion>1</MinorVersion>
     <PatchVersion>0</PatchVersion>
     <PackageLifeCycleStage>Development</PackageLifeCycleStage>
     <PackageLabel>Alpha</PackageLabel>
+    <EnableDefaultItems>false</EnableDefaultItems>
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\ATAP.Console.PluginDemo.StringConstants\ATAP.Console.PluginDemo.StringConstants.csproj" />
-    <ProjectReference Include="..\ATAP.Utilities.Configuration\Extensions\ATAP.Utilities.Configuration.Extensions.csproj" />
-    <ProjectReference Include="..\ATAP.Utilities.Configuration\Secrets\ATAP.Utilities.Configuration.Secrets.csproj" />
-    <ProjectReference Include="..\ATAP.Utilities.Configuration\Secrets\Shims\Bitwarden\ATAP.Utilities.Configuration.Secrets.Shim.Bitwarden.csproj" />
-    <ProjectReference Include="..\ATAP.Utilities.ETW\ATAP.Utilities.ETW.csproj" />
-    <ProjectReference Include="..\ATAP.Utilities.Loader.Interfaces\ATAP.Utilities.Loader.Interfaces.csproj" />
-    <ProjectReference Include="..\ATAP.Utilities.Loader\ATAP.Utilities.Loader.csproj" />
-    <ProjectReference Include="..\ATAP.Utilities.Plugin\ATAP.Utilities.Plugin.csproj" />
+    <ProjectReference Include="Model\ATAP.Console.PluginDemo.Model.csproj" />
+    <ProjectReference Include="StringConstants\ATAP.Console.PluginDemo.StringConstants.csproj" />
   </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.Hosting" />
-    <PackageReference Include="Microsoft.Extensions.Logging" />
-    <PackageReference Include="Serilog.Extensions.Hosting" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <None Update="appsettings.json">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
-    <None Update="appsettings.Development.json">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
-  </ItemGroup>
-
-  <Target Name="CopyPluginToOutputPluginsDirectory" AfterTargets="Build">
-    <MakeDir Directories="$(OutputPath)Plugins" />
-    <Copy
-      SourceFiles="$(OutputPath)ATAP.Utilities.Configuration.Secrets.Shim.Bitwarden.dll"
-      DestinationFolder="$(OutputPath)Plugins"
-      SkipUnchangedFiles="true"
-      ContinueOnError="WarnAndContinue" />
-    <Copy
-      SourceFiles="$(OutputPath)ATAP.Utilities.Configuration.Secrets.dll"
-      DestinationFolder="$(OutputPath)Plugins"
-      SkipUnchangedFiles="true"
-      ContinueOnError="WarnAndContinue" />
-  </Target>
 
 </Project>
