+++ src/ATAP.Console.PluginDemo/Model/PluginDemoBackgroundService.cs
