+++ src/ATAP.Console.PluginDemo/Model/PluginDemoDefaultConfiguration.cs
