+++ src/ATAP.Console.PluginDemo/Model/Program.cs
