+++ src/ATAP.Console.PluginDemo/Model/Properties/AssemblyInfo.cs
