+++ src/ATAP.Console.PluginDemo/Model/appsettings.json
