+++ src/ATAP.Console.PluginDemo/StringConstants/Properties/AssemblyInfo.cs
@@ -1,0 +1,5 @@
+using System.Reflection;
+
+[assembly: AssemblyFileVersion("0.1.0.0")]
+[assembly: AssemblyInformationalVersion("0.1.0")]
+[assembly: AssemblyVersion("0.1.0")]
