+++ src/ATAP.Console.PluginDemo/StringConstants/StringConstants.cs
