+++ src/ATAP.Utilities.Loader/Interfaces/ATAP.Utilities.Loader.Interfaces.csproj
@@ -23,7 +23,7 @@
 
   <!-- Extensions methods on File and Directory types, along with a Globbing library -->
   <ItemGroup>
-    <ProjectReference Include="..\ATAP.Utilities.FIleIO\ATAP.Utilities.FileIO.csproj" />
+    <ProjectReference Include="..\..\ATAP.Utilities.FIleIO\ATAP.Utilities.FileIO.csproj" />
   </ItemGroup>
 
 </Project>
