+++ src/ATAP.Utilities.Loader/Interfaces/ILoader.cs
