+++ src/ATAP.Utilities.Loader/Interfaces/ILoaderOptions.cs
