+++ src/ATAP.Utilities.Loader/Interfaces/Properties/AssemblyInfo.cs
