+++ src/ATAP.Utilities.Loader/Model/CollectibleAssemblyLoadContext.cs
