+++ src/ATAP.Utilities.Loader/Model/LoadedAssemblyContext.cs
