+++ src/ATAP.Utilities.Loader/Model/LoaderOptions.cs
