+++ src/ATAP.Utilities.Loader/Model/NativeAssemblyLoader.cs
