+++ src/ATAP.Utilities.Loader/StringConstants/Properties/AssemblyInfo.cs
