+++ src/ATAP.Utilities.Loader/StringConstants/Properties/PublishProfiles/Release.win-x64.SingleFile.pubxml
