+++ src/ATAP.Utilities.Loader/StringConstants/Properties/launchSettings.json
