+++ src/ATAP.Utilities.MessageQueue/Shim/RabbitMQ/ATAP.Utilities.MessageQueue.Shim.RabbitMQ.csproj
@@ -49,7 +49,6 @@
 
   <!-- Loader interfaces and classes for any Assembly that may be loaded by the dynamic Loader, only works for TargetFramework NetCore -->
   <ItemGroup Condition="$(TargetFramework.Contains('netcore'))">
-    <ProjectReference Include="..\..\ATAP.Utilities.Loader.Interfaces\ATAP.Utilities.Loader.Interfaces.csproj" />
     <ProjectReference Include="..\..\ATAP.Utilities.Loader\ATAP.Utilities.Loader.csproj" />
   </ItemGroup>
