+++ src/ATAP.Utilities.MessageQueue/Shim/TPL/ATAP.Utilities.MessageQueue.Shim.TPL.csproj
@@ -53,10 +53,9 @@
     <PackageReference Include="Serilog" />
   </ItemGroup>
 
-  <!-- Loader interfaces and classes for any Assembly that may be loaded by the dynamic Loader, only works for TargetFramework NetCore
+  <!-- Loader interfaces and classes for any Assembly that may be loaded by the dynamic Loader, only works for TargetFramework NetCore -->
   <ItemGroup Condition="$(TargetFramework.Contains('netcore'))">
-    <ProjectReference Include="..\ATAP.Utilities.Loader.Interfaces\ATAP.Utilities.Loader.Interfaces.csproj" />
     <ProjectReference Include="..\ATAP.Utilities.Loader\ATAP.Utilities.Loader.csproj" />
   </ItemGroup>
- -->
+
 </Project>
