+++ src/ATAP.Utilities.Secrets/Shim/Plugin/ATAP.Utilities.Secrets.Shim.Plugin.csproj
@@ -14,7 +14,6 @@
     <ProjectReference Include="..\..\Interfaces\ATAP.Utilities.Secrets.Interfaces.csproj" />
     <ProjectReference Include="..\..\StringConstants\ATAP.Utilities.Secrets.StringConstants.csproj" />
     <ProjectReference Include="..\..\..\ATAP.Utilities.Plugin\Model\ATAP.Utilities.Plugin.Model.csproj" />
-    <ProjectReference Include="..\..\..\ATAP.Utilities.Loader.Interfaces\ATAP.Utilities.Loader.Interfaces.csproj" />
     <ProjectReference Include="..\..\..\ATAP.Utilities.Loader\ATAP.Utilities.Loader.csproj" />
   </ItemGroup>
