+++ src/ATAP.Utilities.Serializer/Shim/Plugin/ATAP.Utilities.Serializer.Shim.Plugin.csproj
@@ -38,7 +38,6 @@
 
   <!-- Loader interfaces and classes for any Assembly that may be loaded by the dynamic Loader, only works for TargetFramework NetCore -->
   <ItemGroup Condition="$(TargetFramework.Contains('netcore'))">
-    <ProjectReference Include="..\..\..\ATAP.Utilities.Loader.Interfaces\ATAP.Utilities.Loader.Interfaces.csproj" />
     <ProjectReference Include="..\..\..\ATAP.Utilities.Loader\ATAP.Utilities.Loader.csproj" />
   </ItemGroup>
