+++ src/ATAP.Utilities.Testing.Fixture.Serialization.Shim.Plugin/ATAP.Utilities.Testing.Fixture.Serialization.Shim.Plugin.csproj
@@ -103,7 +103,6 @@
 
   <!-- Loader interfaces and classes for any Assembly that may be loaded by the dynamic Loader, only works for TargetFramework NetCore -->
   <ItemGroup Condition="$(TargetFramework.Contains('5.0'))">
-    <ProjectReference Include="..\ATAP.Utilities.Loader.Interfaces\ATAP.Utilities.Loader.Interfaces.csproj" />
     <ProjectReference Include="..\ATAP.Utilities.Loader\ATAP.Utilities.Loader.csproj" />
   </ItemGroup>
